service added

+++ Core/Entities/CustomerBasket.cs
@@ -8,7 +8,7 @@ public class CustomerBasket
     public string ClientSecret { get; set; }
     public string PaymentIntentId { get; set; }
     public string DeliveryMethod { get; set; }
-    public decimal ShippingPrice { get; set; }
+    public double ShippingPrice { get; set; }
     public CustomerBasket() { }
     public CustomerBasket(string id) => Id = id;
 }
